rxtools/source/lib: Add lang.c
